+++ crates/graphql-composition/src/subgraphs.rs
@@ -192,4 +192,15 @@ impl Subgraphs {
     pub(crate) fn emit_ingestion_diagnostics(&self, diagnostics: &mut crate::Diagnostics) {
         diagnostics.clone_all_from(&self.ingestion_diagnostics);
     }
+
+    /// After subgraphs have been ingested, we have to sort some of the vecs we expect to be sorted at the composition stage, because with type extensions, they may be out of order. For example keys may have had other definitions ingested before we are done ingesting a given type (always because of type extensions).
+    pub(crate) fn sort_post_ingestion(&mut self) {
+        self.keys.keys.sort_unstable_by_key(|key| key.definition_id);
+        self.directives
+            .extra_directives
+            .sort_unstable_by_key(|directive| directive.directive_site_id);
+        self.fields
+            .fields
+            .sort_unstable_by_key(|field| field.parent_definition_id);
+    }
 }