+++ gateway/tests/entity_caching/mod.rs
@@ -63,7 +63,7 @@ where
         subgraphs
             .ingest_str(subgraph_schema, "the-subgraph", Some(subgraph_url))
             .unwrap();
-        let graph = graphql_composition::compose(&subgraphs).into_result().unwrap();
+        let graph = graphql_composition::compose(&mut subgraphs).into_result().unwrap();
         graphql_composition::render_federated_sdl(&graph).unwrap()
     };